requirements.txt
amqp==2.6.1
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
billiard==3.6.3.0
bleach==3.2.0
celery==4.4.7
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.2
Flask-Cors==3.0.9
Flask-SQLAlchemy==2.4.4
flower==0.9.5
gevent==20.9.0
greenlet==0.4.17
gunicorn==20.0.4
humanize==2.6.0
icc-rt==2020.0.133
idna==2.10
importlib-metadata==1.7.0
intel-numpy==1.15.1
intel-openmp==2020.0.133
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.14.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.1
kiwisolver==1.2.0
kombu==4.6.11
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
mkl==2019.0
mkl-fft==1.0.6
mkl-random==1.0.1.1
nbclient==0.5.0
nbconvert==6.0.3
nbformat==5.0.7
nest-asyncio==1.4.0
notebook==6.1.4
numpy==1.19.4
olefile==0.46
open3d @ file:///home/yma50/development/O3dInstall/Open3D/build/lib/python_package/pip_package/open3d-0.11.1%2Bb6cf739-cp36-cp36m-linux_x86_64.whl
opencv-python==4.4.0.44
package-name==0.1
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
Paste==3.4.4
PasteDeploy==2.1.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.13.0
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytimeparse==1.1.8
pytz==2020.1
pyzmq==19.0.2
redis==3.5.3
requests==2.24.0
Send2Trash==1.5.0
six==1.15.0
SQLAlchemy==1.3.19
tbb==2019.0
tbb4py==2019.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.10
vine==1.3.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
zipp==3.1.0
zope.event==4.5.0
zope.interface==5.1.2
